Backend/src/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
PyJWT==2.8.0
PyMySQL==1.0.2
python-dotenv==1.0.0
Werkzeug==2.2.2
